Adicionar a regra para só os diretores e gerentes poderem excluir os projetos

+++ src/main/java/br/com/zup/sistemareembolso/services/ProjetoService.java
@@ -72,7 +72,13 @@ public class ProjetoService {
         throw new ProjetoNaoExistenteException();
     }
 
-    public void excluirProjetoPeloId(int id) {
+    public void excluirProjetoPeloId(int id, Colaborador colaborador) {
+        Colaborador colaboradorDoBanco = colaboradorService.pesquisarColaboradorPorCpf(colaborador.getCpf());
+
+        if (!colaborador.getCargo().equals(Cargo.GERENTE) && !colaborador.getCargo().equals(Cargo.DIRETOR)) {
+            throw new PermissaoNegadaParaExcluirProjetoException();
+        }
+
         Projeto projeto = pesquisarProjetoPeloId(id);
 
         projetoRepository.delete(projeto);
